Wallet module reorg
