--- sdk/core/src/crypto.rs
@@ -1,3 +1,0 @@
-pub mod private_key;
-pub mod public_key;
-pub mod wallet_signature;
