+++ sdk/core/src/lib.rs
@@ -3,7 +3,6 @@
 #![allow(clippy::collapsible_if)]
 #![allow(clippy::manual_is_multiple_of)]
 
-pub mod crypto;
 pub mod data;
 pub mod gateway;
 pub mod retrieve_tx_on_network;
