+++ sdk/core/src/wallet/wallet_signature.rs
